Theo/development

+++ sql/functions/format_parcelId.sql
@@ -9,9 +9,10 @@ $BODY$
 DECLARE
 	ParcelId varchar;
 BEGIN
-
+-- Old version of this function before they switched to the building all table.  (Parcel was an int and is now a float in the BldgSect table)
 --ParcelId := replace(replace(concat(to_char(CityBlock::float8,'0000.00'),to_char(Parcel::int8,'0000'),OwnerCode),'.',''),' ','');
-ParcelId := (concat(to_char(CityBlock::float8,'0000.00'),to_char(Parcel::int8,'0000'),OwnerCode));
+
+ParcelId := (concat(to_char(CityBlock::float8,'0000.00'),to_char(Parcel::float8,'0000'),OwnerCode));
 ParcelId := replace((ParcelId), '.', '');
 ParcelId := replace((ParcelId), ' ', '');
